tasasweb/requirements.txt
app-version==0.2.1
appdirs==1.4.3
attrs==19.1.0
backports.shutil-get-terminal-size==1.0.0
cached-property==1.3.1
cffi==1.1.2
characteristic==14.3.0
chardet==2.3.0
click==7.1.1
colorama==0.3.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==1.0.1
cssselect==0.9.1
defusedxml==0.5.0
Django==1.11.23
django-ajax-selects==1.5.0
django-appconf==1.0.2
django-bootstrap3==11.0.0
django-chartit2==0.2.2
django-chosen==0.1
django-compressor==2.2
django-cors-headers==3.0.2
django-crispy-forms==1.5.0
django-debug-toolbar==1.11
django-dowser==0.1.4
django-easy-pdf==0.1.1
django-extensions==2.2.9
django-extra-views==0.7.1
django-firebird==1.8
django-fm==0.2.5
django-geoposition==0.3.0
django-localflavor==1.5.2
django-material==0.8.0
django-rest==0.0.1
django-rest-framework==0.1.0
django-rest-framework-docs==0.1.7
django-rest-swagger==2.2.0
django-secure==1.0.1
django-session-security==2.6.0
django-suit==0.2.18
django-widget-tweaks==1.4.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
docopt==0.6.2
Dozer==0.5
drf-yasg==1.16.1
enum34==1.1.4
fdb==2.0.1
fpdf==1.7.2
funcsigs==1.0.2
future==0.18.2
futures==3.2.0
GRR-M2Crypto==0.22.6.post2
html5lib==1.0.1
httplib2==0.9.2
idna==2.1
image==1.5.27
inflection==0.3.1
ipaddress==1.0.16
isodate==0.5.4
itypes==1.1.0
Jinja2==2.10.1
jsonpickle==1.2
lxml==3.4.4
Mako==1.0.3
MarkupSafe==0.23
mock==3.0.5
model-mommy==2.0.0
mysqlclient==1.3.7
openapi-codec==1.3.2
packaging==19.1
pbr==5.2.1
Pillow==6.0.0
pipreqs==0.4.2
pyasn1==0.1.9
pycparser==2.14
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.2
PyPDF2==1.25
Pyphen==0.9.1
PySimpleSOAP==1.16
python-dateutil==2.5.3
python-slugify==1.2.0
pytz==2017.2
rcssmin==1.0.6
reportlab==3.2.0
requests==2.10.0
requests-toolbelt==0.8.0
rjsmin==1.0.12
ruamel.ordereddict==0.4.14
ruamel.yaml==0.16.1
ruamel.yaml.clib==0.1.2
selenium==3.141.0
simplejson==3.13.2
six==1.10.0
soap2py==1.16
SOAPpy==0.12.22
sqlparse==0.3.0
stevedore==1.30.1
suds-jurko==0.6
tinycss==0.3
tolerance==0.1.2
typing==3.7.4.1
Unidecode==0.4.19
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==13.1.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.5
webassets==0.12.1
webencodings==0.5.1
WebOb==1.5.1
Werkzeug==0.14.1
wstools==0.4.5
xhtml2pdf==0.2.3
xlrd==0.9.4
xlutils==1.7.1
xlwt==1.0.0
yarg==0.1.9
zeep==2.4.0
